+++ fregl/fregl_roi.h
@@ -7,7 +7,7 @@
 class fregl_roi
 {
 public:
-	typedef itk::Image< unsigned int, 3 >  ImageType;
+	typedef itk::Image< unsigned short, 3 >  ImageType;
 
 	fregl_roi(std::string joint_xforms_xml_file, std::string img_path, std::string anchor_image, bool nearest_neighbor = false);
 	~fregl_roi();
